requirements.txt
asgiref==3.3.4
Django==3.2
django-debug-toolbar==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
psycopg2==2.8.6
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
